requirements.txt
comet_ml
load_dotenv
numpy
pandas
pydantic
matplotlib
seaborn
requests
tqdm
plotly
ipywidgets
ipykernel
jupyterlab>=3.0.0
setuptools